Implement queue.delete, exchange.delete and handling of respective .delete-ok methods
 * Note that we currently do not clear delete_* lists on channel or connection exceptions

 * I also removed immediate callback execution for nowait case because it is not clear how
   this can be useful.

So why callback execution with nowait is an arguable feature?

First of all, since there is no reply from the broker, we canno
yield any parameters to this callback. So callback would have to handle two different
parameter sets and it violates single responsibility principle severly.

Secondly, that callback will be executed even before method returns, this is a case of
an unusual sync flow and is very counter-intuitive. If you want to do something right after
you call Queue#declare (or something like this) with nowait, just do it after that method
returns.

So for now, it is not worth having. However, amqp-gem 0.7 may work that way :( So chances are,
this broken behavior will be back. Indeed, most bad decisions tend to stay around for a
long time.