Towards working basic.ack and basic.reject (read on)
Since we need to make it possible for delivery handler to acknowledge or reject
messages (and specify parameters like requeueing along the way), a decision has to
be made about how exactly to do this.

After playing with a couple of solutions
that relied on return values from the block, I decided to simply provide Queue#acknowledge
and Queue#reject. Since Queue instance is passed on to delivery handler, end application
can do anything it wants with messages. Simple, straightforward, infinitely flexible
and easy to test. Win.