Do not yield self to callbacks (read on)
AMQP gem porting reveals that most of higher-level libraries that
use callbacks don't need this initial argument of self. This is primarily because
it is present in the lexical scope of the block.

Without this change we have to use shim functions and it both harms
performance of a few important methods (like basic.delivery handler) and
raises complexity. All while not really contributing anything: it just
bridges two APIs together.

Alternative would be to override #handle_* methods so that they use
a version of Entity#exec_callback that does yield self.