Connection loss detection is hard, lets go shopping
Before EventMachine 1.0.beta3 or so #unbind serves at least 4
purposes with no clear indication of why exactly it was called. This makes
connection loss detection a little challenging.

There is no reconnection logic built into amq-client, by design. All we guarantee
is that when connection is lost, we call X. When it fails the 1st time,
we call Y but not X. And when reconnection happens, we go through AMQP
connection negotiation process again.