Decouple Consumer from Queue, enables several important aspects of functionality
First of all, no more "one consumer per queue" limitation. Not only there is no
reason for it, but also message load balancing happens on the consumer level.
This forces amqp gem users today to use multiple channels.

To preserve backwards compatibility, Queue#consume is not going anywhere, just
relies on the default consumer, of which every Queue still has one. This is also
very similar to the way the RabbitMQ Java client works, although default consumer
there has a few special cases we do not implement here.

Another upside of this decoupling is that auto-recovery is now possible. Consumer
is an entity that encapsulates subscription metadata (consumer tag, exclusivity and so on)
and can use it on recovery. Ruby blocks, however, do not store any information
and thus recovery for them is not possible. Ability to recover queue consumers
is the key feature for automatic recovery for consumer apps.

A very important step forward for amq-client.