0.8.0 (read on)
We want to stop using pre-release dependencies for the amqp gem and
align versions of amq-client and amq-protocol with the amqp when possible.

The primary reason for doing so is that RubyGems 1.3.7, also known as
'the Internet Explorer 6 of the Ruby ecosystem', has buggy pre-release
version comparison logic: it considers 0.8.0.rc9 to be more recent than
0.8.0.rc13, for example. While many developers & companies have moved
beyond 1.3.7 many still did not, and apps hosted on Heroku still use
RubyGems 1.3.7. So lets stop using pre-release dependencies instead.
And this entails synchronized gem releases, one way or another.

amq-protocol 0.8.0 has been released for multiple days already. So lets
go to 0.8.0 for amq-client, too.